requirements.txt
validate-email~=1.3
PyYAML~=3.12
